requirements.txt
Pillow==10.2.0
numpy==1.26.4
beautifulsoup4==4.12.3
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.27.1
cairosvg==2.7.1 